refactor: style.ts에서 cn.ts로 네이밍 수정
